Fix sporadic deadlock issue
The program was sometimes freezing because no mutex was guarding the
check of the clerk status when entering the request_service function. We
need to lock it here, and release the lock after setting the the clerk
to busy when a customer is served. We then need to re-acquire when we
set the clerk back to idle.